+++ src/main/java/webserver/controller/Controller.java
@@ -1,0 +1,8 @@
+package webserver.controller;
+
+import webserver.http.HttpRequest;
+import webserver.http.HttpResponse;
+
+public interface Controller {
+    HttpResponse service(HttpRequest httpRequest);
+}
